requirements.txt
certifi==2025.7.14
charset-normalizer==3.4.2
distlib==0.4.0
docutils==0.21.2
filelock==3.18.0
filetype==1.2.0
idna==3.10
Kivy==2.3.1
Kivy-examples==2.3.1
Kivy-Garden==0.1.5
kivymd==1.2.0
pdf2image==1.17.0
pillow==10.4.0
platformdirs==4.3.8
plyer==2.1.0
Pygments==2.19.2
PyPDF2==3.0.1
requests==2.32.4
setuptools==80.9.0
urllib3==2.5.0
virtualenv==20.31.2
